requirements.txt
streamlit
psycopg2-binary
python-dotenv
requests
beautifulsoup4
generativedai